% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFunctions.R
\name{getSummaryData}
\alias{getSummaryData}
\title{getSummaryData}
\usage{
getSummaryData(rawData)
}
\arguments{
\item{rawData}{A data.frame}
}
\value{
the first n rows of the data.frame, where n is determined by the
R_CONFIG_ACTIVE (config package)
}
\description{
Placeholder function for data processing
}
